% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fam_info_w3.R
\docType{data}
\name{fam_info_w3}
\alias{fam_info_w3}
\title{Information at family level.}
\format{
A data frame with 2720 rows and 2 variables:
\describe{
\item{OTUID}{ID of OTU}
\item{Family}{Family name}
}
}
\source{
\url{https://github.com/N-T-Huyen}
}
\usage{
data(fam_info_w3)
}
\description{
A dataset containing the information at family level.
}
\keyword{datasets}
