% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fam_shan_trim_w3.R
\docType{data}
\name{fam_shan_trim_w3}
\alias{fam_shan_trim_w3}
\title{Dataset at family level.}
\format{
A data frame with 6 rows and 82 variables:
\describe{
Rows are family names and columns are names of subjects.
}
}
\source{
\url{https://github.com/N-T-Huyen}
}
\usage{
data(fam_shan_trim_w3)
}
\description{
A dataset containing the Shannon index of 6 families after filtering.
}
\keyword{datasets}
