% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetRA.R
\name{GetRA}
\alias{GetRA}
\title{This function convert OTU matrix to RA matrix.}
\usage{
GetRA(Micro.mat)
}
\arguments{
\item{Micro.mat}{an OTU matrix with OTUs in rows and subjects in columns.}
}
\value{
A relative abundance matrix of OTUs
\item{ra}{Relative abundance matrixs}
}
\description{
This function convert OTU matrix to RA matrix.
}
\examples{
# Read dataset
data(Week3_otu)
Week3_otu = data.frame(Week3_otu)
otu_mat_w3 = t(data.matrix(Week3_otu[ , 1:2720]))

# Convert absolute abundance to relative abundance
ra_otu_trim_w3 = GetRA(Micro.mat = otu_mat_w3)
}
\seealso{
\code{\link[MicrobiomeSurv]{GetRA}}
}
\author{
Thi Huyen Nguyen, \email{thihuyen.nguyen@uhasselt.be}

Olajumoke Evangelina Owokotomo, \email{olajumoke.x.owokotomo@gsk.com}

Ziv Shkedy
}
