% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_taxonomy.R
\docType{data}
\name{metadata_taxonomy}
\alias{metadata_taxonomy}
\title{Metadata taxonomy.}
\format{
A data frame with 2720 rows and 3 variables:
\describe{
\item{OTUID,Taxon,Confidence}{OTU ID and information at higher levels}
  ...
}
}
\source{
\url{https://elifesciences.org/articles/37816}
}
\usage{
data(metadata_taxonomy)
}
\description{
A dataset containing the information of all levels in the ecosystem: OTU, order, family, kingdom, ...
}
\keyword{datasets}
