% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Week3_otu.R
\docType{data}
\name{Week3_otu}
\alias{Week3_otu}
\title{OTU table at week 3.}
\format{
A data frame with 81 rows and 2724 variables, we only use 2720 first variables:
\describe{
\item{X226097bd7a1661a286a3b62d1c1f0e3a-X005d3193f381b0793f0c928bde66dd21}{Names of the OTUs}
\item{SampleID}{ID of the subject}
\item{Treatment}{Treatment variable}
\item{T1DWeek}{Time to develop T1D in week}
\item{T1D}{Censored indicator}
}
}
\source{
\url{https://elifesciences.org/articles/37816}
}
\usage{
data(Week3_otu)
}
\description{
A dataset containing the count of OTUs.
}
\keyword{datasets}
