% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Week3_response.R
\docType{data}
\name{Week3_response}
\alias{Week3_response}
\title{Response datase.}
\format{
A data frame with 81 rows and 30 variables:
\describe{
\item{SampleID}{ID of the subject}
\item{Treatment}{Treatment variable}
\item{T1Dweek}{Time to develop T1D in week}
\item{T1D}{Censored indicator}
\item{Treatment_new}{Treatment indicator obtained from treatment variable}
}
}
\source{
\url{https://elifesciences.org/articles/37816}
}
\usage{
data(Week3_response)
}
\description{
A dataset containing the information of subjects.
}
\keyword{datasets}
