% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZerosPerGroup.R
\name{ZerosPerGroup}
\alias{ZerosPerGroup}
\title{This function returns a matrix with rows are Micros and 9 columns containing number and the proportion of zeros per groups of treatments and in total.}
\usage{
ZerosPerGroup(
  Micro.mat,
  groups,
  week = 0,
  n.obs = n.obs,
  n.control = n.control,
  n.treated = n.treated,
  n.mi = n.mi,
  plot = FALSE
)
}
\arguments{
\item{Micro.mat}{Micro matrix (rows are Micros, columns are subjects)}

\item{groups}{Treatment groups or groups of any binary variables}

\item{week}{A specific time point. To use when having different time points in the dataset.}

\item{n.obs}{Number of patients.}

\item{n.control}{Number of patients in control group or in the first group.}

\item{n.treated}{Number of patients in treated group or in the second group.}

\item{n.mi}{Number of taxa.}

\item{plot}{A boolean parameter indicating if the plot should be shown. Default is FALSE}
}
\value{
A matrix with information of number and the proportion of zeros per groups.
\item{zero.per.group}{A matrix with rows are Micros and 9 columns containing number and the proportion of zeros per groups of treatments and in total.}
\item{plot}{Plot percentage of zeros per group}
}
\description{
This function returns a matrix with rows are Micros and 9 columns containing number and the proportion of zeros per groups of treatments and in total.
}
\seealso{
\code{\link[MicrobiomeSurv]{ZerosPerGroup}}
}
\author{
Thi Huyen Nguyen, \email{thihuyen.nguyen@uhasselt.be}

Olajumoke Evangelina Owokotomo, \email{olajumoke.x.owokotomo@gsk.com}

Ziv Shkedy
}
