% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Majorityvotes.R
\name{Majorityvotes}
\alias{Majorityvotes}
\title{Classifiction for Majority Votes}
\usage{
Majorityvotes(Result, Prognostic, Survival, Censor, J = 1)
}
\arguments{
\item{Result}{An object obtained from the taxon specific analysis (\code{\link[MicrobiomeSurv]{MSpecificCoxPh}}) which is of class "ms"}

\item{Prognostic}{A dataframe containing possible prognostic(s) factor and/or treatment effect to be used in the model.}

\item{Survival}{A vector of survival time with length equals to number of subjects}

\item{Censor}{A vector of censoring indicator}

\item{J}{The jth set of subjects required for the visualization. The default is J=1 which is the first set of subjects. For visualization, J should be less than the number of subjects divided by 25}
}
\value{
A list is returned with the following values
  \item{Model.result}{The cox proportional regression result based on the majority vote classification}
  \item{N}{The majority vote for each subject}
  \item{Classif}{The majority vote classification for each subjects}
  \item{Group}{The classification of the subjects based on each taxon analysis}
}
\description{
The Function fits cox proportional hazard model and does classification based on the majority votes.
}
\details{
The Function fits cox proportional hazard model and does classification based on the majority votes
while estimating the Hazard ratio of the low risk group.
The function firstly count the number of low risk classification for each subject based on the taxon specific analysis
which determines the majority votes.
In addition, function visualizes the taxon specific calssification for the subjects. 25 subjects is taken for visualization purpose.
}
\seealso{
\code{\link[MicrobiomeSurv]{MSpecificCoxPh}}, \code{\link[survival]{coxph}},  \code{\link[MicrobiomeSurv]{EstimateHR}}
}
\author{
Thi Huyen Nguyen, \email{thihuyen.nguyen@uhasselt.be}

Olajumoke Evangelina Owokotomo, \email{olajumoke.x.owokotomo@gsk.com}

Ziv Shkedy
}
