% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiFreq.R
\name{MiFreq}
\alias{MiFreq}
\title{Frequency of Selected Taxa from the LASSO, Elastic-net Cross-Validation}
\usage{
MiFreq(Object, TopK = 20)
}
\arguments{
\item{Object}{An object of class \code{\link[MicrobiomeSurv]{cvle}} returned from the function \code{\link[MicrobiomeSurv]{CVLasoelacox}}.}

\item{TopK}{The number of Top K taxa (5 by default) to be displayed in the frequency of selection graph.}
}
\value{
A vector of taxa and their frequency of selection. Also, a graphical representation is displayed.
}
\description{
The function selects the frequency of selection from the shrinkage method (LASSO, Elastic-net) based on cross validation,
that is the number of times each taxon occur during the cross-validation process.
This function outputs the mostly selected taxa during the LASSO and Elastic-net cross validation.
Selected top taxa are ranked based on frequency of selection and also a particular frequency can be selected.
In addition, it visualizes the selected top taxa based on the minimum frequency specified.
}
\seealso{
\code{\link[MicrobiomeSurv]{cvmm}}, \code{\link[survival]{coxph}},
\code{\link[MicrobiomeSurv]{EstimateHR}},\code{\link[MicrobiomeSurv]{CVLasoelacox}}
}
\author{
Thi Huyen Nguyen, \email{thihuyen.nguyen@uhasselt.be}

Olajumoke Evangelina Owokotomo, \email{olajumoke.x.owokotomo@gsk.com}

Ziv Shkedy
}
