% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryData.R
\name{SummaryData}
\alias{SummaryData}
\title{This function gives indices such as Observed richness, Shannon index, Inverse Simpson, ... of higher level such as levelily, order, phylum, ...}
\usage{
SummaryData(Micro.mat, info, measure = "observed")
}
\arguments{
\item{Micro.mat}{an OTU matrix with OTUs in rows and subjects in columns.}

\item{info}{A n x 2 matrix containing a column of OTU's names and a column of the corresponding information of the chosen level.}

\item{measure}{The indices at chosen level that user wishes to use. It can be observed richness, Shannon index, inverse Simpson, ...}
}
\value{
A matrix of the selected measurement of the chosen level.
\item{level.measure}{A matrix of measurements at levelily level of patients}
}
\description{
This function gives indices such as Observed richness, Shannon index, Inverse Simpson, ... of higher level such as levelily, order, phylum, ...
}
\examples{
\donttest{
# Preparing data for analysis at family level
fam = otus_data[c("Family", "OTUID")]
otus$OTUID = row.names(otus)
otufam = inner_join(otus, fam)
otu_fam_w3 = otufam[ ,which(colnames(otufam) \%in\% c(Week3_response$SampleID, "OTUID", "Family"))]
fam_w3 = otu_fam_w3[ ,c("OTUID", "Family")]
otu_data_w3 = otu_fam_w3[ , 1:n_obs]
rownames(otu_data_w3) = otu_fam_w3$OTUID

# USing the function
fam_shan_w3 = SummaryData(Micro.mat = otu_data_w3, info = fam_w3, measure = "shannon")
}
}
\seealso{
\code{\link[MicrobiomeSurv]{SummaryData}}
}
\author{
Thi Huyen Nguyen, \email{thihuyen.nguyen@uhasselt.be}

Olajumoke Evangelina Owokotomo, \email{olajumoke.x.owokotomo@gsk.com}

Ziv Shkedy
}
