% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryData.R
\name{SummaryData}
\alias{SummaryData}
\title{This function gives indices such as Observed richness, Shannon index, Inverse Simpson, ... of higher level such as levelily, order, phylum, ...}
\usage{
SummaryData(Micro.mat, info, measure = "observed")
}
\arguments{
\item{Micro.mat}{an OTU matrix with OTUs in rows and subjects in columns.}

\item{info}{A n x 2 matrix containing a column of OTU's names and a column of the corresponding information of the chosen level.}

\item{measure}{The indices at chosen level that user wishes to use. It can be observed richness, Shannon index, inverse Simpson, ...}
}
\value{
A matrix of the selected measurement of the chosen level.
\item{level.measure}{A matrix of measurements at levelily level of patients}
}
\description{
This function gives indices such as Observed richness, Shannon index, Inverse Simpson, ... of higher level such as levelily, order, phylum, ...
}
\seealso{
\code{\link[MicrobiomeSurv]{SummaryData}}
}
\author{
Thi Huyen Nguyen, \email{thihuyen.nguyen@uhasselt.be}

Olajumoke Evangelina Owokotomo, \email{olajumoke.x.owokotomo@gsk.com}

Ziv Shkedy
}
